Update Cluster Storage Connected Resources and Delete Modal (#4728)
* Update connected resources and delete modal

* update comment

* add better bullet spacing
