+++ frontend/src/pages/projects/pvc/DeletePVCModal.tsx
@@ -6,8 +6,9 @@ import {
   ConnectedNotebookContext,
 } from '#~/pages/projects/notebook/useRelatedNotebooks';
 import DeleteModal from '#~/pages/projects/components/DeleteModal';
-import DeleteModalConnectedAlert from '#~/pages/projects/components/DeleteModalConnectedAlert';
+import ConnectedResourcesDeleteModal from '#~/pages/projects/components/ConnectedResourcesDeleteModal';
 import { getDisplayNameFromK8sResource } from '#~/concepts/k8s/utils';
+import { useInferenceServicesForConnection } from '#~/pages/projects/useInferenceServicesForConnection';
 
 type DeletePVCModalProps = {
   pvcToDelete: PersistentVolumeClaimKind;
@@ -17,11 +18,11 @@ type DeletePVCModalProps = {
 const DeletePVCModal: React.FC<DeletePVCModalProps> = ({ pvcToDelete, onClose }) => {
   const [isDeleting, setIsDeleting] = React.useState(false);
   const [error, setError] = React.useState<Error | undefined>();
-  const {
-    notebooks: connectedNotebooks,
-    loaded: notebookLoaded,
-    error: notebookError,
-  } = useRelatedNotebooks(ConnectedNotebookContext.EXISTING_PVC, pvcToDelete.metadata.name);
+  const { notebooks: connectedNotebooks, loaded: notebookLoaded } = useRelatedNotebooks(
+    ConnectedNotebookContext.EXISTING_PVC,
+    pvcToDelete.metadata.name,
+  );
+  const connectedModels = useInferenceServicesForConnection(pvcToDelete);
 
   const onBeforeClose = (deleted: boolean) => {
     onClose(deleted);
@@ -58,10 +59,13 @@ const DeletePVCModal: React.FC<DeletePVCModalProps> = ({ pvcToDelete, onClose })
       error={error}
       deleteName={displayName}
     >
-      <DeleteModalConnectedAlert
+      The <b>{getDisplayNameFromK8sResource(pvcToDelete)}</b> storage will be deleted, and its
+      dependent resources will stop working.
+      <ConnectedResourcesDeleteModal
         connectedNotebooks={connectedNotebooks}
-        error={notebookError}
+        connectedModels={connectedModels}
         loaded={notebookLoaded}
+        namespace={pvcToDelete.metadata.namespace}
       />
     </DeleteModal>
   );
