+++ frontend/src/pages/projects/useInferenceServicesForConnection.ts
@@ -1,10 +1,11 @@
 import * as React from 'react';
-import { InferenceServiceKind } from '#~/k8sTypes';
+import { InferenceServiceKind, PersistentVolumeClaimKind } from '#~/k8sTypes';
 import { Connection } from '#~/concepts/connectionTypes/types';
 import { ProjectDetailsContext } from '#~/pages/projects/ProjectDetailsContext';
+import { getPVCNameFromURI } from '#~/pages/modelServing/screens/projects/utils';
 
 export const useInferenceServicesForConnection = (
-  connection?: Connection,
+  connection?: Connection | PersistentVolumeClaimKind,
 ): InferenceServiceKind[] => {
   const {
     inferenceServices: {
@@ -18,8 +19,10 @@ export const useInferenceServicesForConnection = (
 
   return inferenceServices.filter(
     (inferenceService) =>
-      // Known issue: this only works for OCI and S3 connections
+      // Known issue: this only works for OCI, S3, and PVC connections
       inferenceService.spec.predictor.model?.storage?.key === connectionName ||
-      inferenceService.spec.predictor.imagePullSecrets?.[0].name === connectionName,
+      inferenceService.spec.predictor.imagePullSecrets?.[0].name === connectionName ||
+      getPVCNameFromURI(inferenceService.spec.predictor.model?.storageUri ?? '') ===
+        connection.metadata.name,
   );
 };
